code refactoring and gemini integrations
