+++ factory/utils/__init__.py
@@ -1,0 +1,9 @@
+"""
+Utility modules for the factory pattern.
+
+This package contains shared utilities that are used across different provider implementations.
+"""
+
+from .file_handler import FileHandler
+
+__all__ = ['FileHandler']
