+++ templates/chat/main.html
@@ -357,6 +357,20 @@
                                                 <span>Grok 4</span>
                                                 <i class="fas fa-check"></i>
                                             </button>
+                                            
+                                            <div class="submenu-group">Google</div>
+                                            <button type="button" class="submenu-option" data-value="gemini_2.5_pro">
+                                                <span>Gemini 2.5 Pro</span>
+                                                <i class="fas fa-check"></i>
+                                            </button>
+                                            <button type="button" class="submenu-option" data-value="gemini_2.5_flash">
+                                                <span>Gemini 2.5 Flash</span>
+                                                <i class="fas fa-check"></i>
+                                            </button>
+                                            <button type="button" class="submenu-option" data-value="gemini_2.5_flash_lite">
+                                                <span>Gemini 2.5 Flash Lite</span>
+                                                <i class="fas fa-check"></i>
+                                            </button>
                                         </div>
                                     </div>
                                 </div>
